* fix entity class `ReplyContentImage` has a wrong primary key * now only save images with 40chars length filename @ `ReplySaver.SavePosts()`
* fix no ignoring updating `IPost.AuthorUid` with `0` since it's only testing against int32 @ `StaticCommonInSavers.FieldChangeIgnorance`
* optimize regex to reduce steps spend on backtracking @ `UserParserAndSaver.PortraitExtractingRegex`
@ crawler

+++ crawler/src/Tieba/Crawl/Saver/StaticCommonInSavers.cs
@@ -60,7 +60,7 @@ public abstract class StaticCommonInSavers
                 && newValue is null && oldValue is not null) return true;
             // possible rarely response with the protoBuf default value 0
             if (propName == nameof(IPost.AuthorUid)
-                && newValue is 0 && oldValue is not null) return true;
+                && newValue is (long)0 && oldValue is not null) return true;
             return false;
         },
         Revision: (whichPostType, propName, oldValue, _) =>
